:truck: Moved json types to json utils
