+++ drawbot/cogs/__init__.py
@@ -1,0 +1,1 @@
+"""Cogs sub-package for drawbot."""
