--- drawbot/json_types.py
@@ -1,5 +1,0 @@
-from typing import Dict, Any, List, Union
-
-JsonDict = Dict[str, Any]
-JsonList = List[Any]
-JsonData = Union[JsonDict, JsonList]
