+++ drawbot/utils/__init__.py
@@ -1,5 +1,11 @@
-from .json_files import json_wr
+from .json_files import json_wr, JsonData, JsonDict, JsonList
 from .pronote import fetch_homeworks, fetch_grades
 
-
-__all__ = ("json_wr", "fetch_homeworks", "fetch_grades")
+__all__ = (
+    "json_wr",
+    "fetch_homeworks",
+    "fetch_grades",
+    "JsonData",
+    "JsonDict",
+    "JsonList",
+)
